Move tests which require GITHUB_TOKEN to IT tests
